hbase: Changed VariantTable rowkey generation. Using binary RK.
This new RowKey uses the same encoding that Phoenix composed rowkeys.
A new table or view can be defined in Phoenix with:

create view "hbase_table_name"  (
        chromosome varchar not null,
        position UNSIGNED_INT not null,
        reference varchar not null ,
        alternate varchar not null
            CONSTRAINT PK PRIMARY KEY (chromosome, position, reference, alternate)
        )
    default_column_family='d';

+++ opencga-storage/opencga-storage-hadoop/src/main/java/org/opencb/opencga/storage/hadoop/variant/archive/ArchiveHelper.java
@@ -200,7 +200,7 @@ public class ArchiveHelper extends GenomeHelper {
 
     public Put wrap (VcfSlice slice) {
 //        byte[] rowId = generateBlockIdAsBytes(slice.getChromosome(), (long) slice.getPosition() + slice.getRecords(0).getRelativeStart() * 100);
-        byte[] rowId = generateBlockIdAsBytes(slice.getChromosome(), (long) slice.getPosition());
+        byte[] rowId = generateBlockIdAsBytes(slice.getChromosome(), slice.getPosition());
         return wrapAsPut(getColumn(), rowId, slice);
     }
